packages/python-sdk/pyproject.toml
[tool.poetry]
name = "nanovm-desktop"
version = "2.0.0"
description = "NanoVM Desktop Sandbox - Deskstop sandbox in cloud powered by NanoVM"
authors = ["nanovm <supriyo@nanovm.dev>"]
license = "MIT"
readme = "README.md"
homepage = "https://app.nanovm.dev/"
repository = "https://github.com/sonite-ai/nanovm-desktop/tree/main/packages/python-sdk"
packages = [{ include = "nanovm_desktop" }]

[tool.poetry.dependencies]
python = "^3.9"

nanovm = "1.0.0"
requests = "^2.32.3"
pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
python-dotenv = "^1.0.0"
pytest-dotenv = "^0.5.2"
pytest-xdist = "^3.6.1"
ruff = "^0.11.12"
pydoc-markdown = "^4.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/sonite-ai/nanovm-desktop/issues"

[tool.ruff.lint]
ignore = ["F401", "F403"]